BASIC - Basic Computer and Internet Policy
This is a general template for a usage policy for a small company. Smaller companies typically tend to have a more lax internet policy, allowing some personal use. While this is not obviously ideal, some companies protect it as part of their culture, and don't want to change these things.